refactoring opinions lists sv to avoid code duplication and to split responsibilities
