+++ backend/opinions-sv/client/src/main/kotlin/com/r8n/backend/opinions/integration/client/OpinionListInternalRestClient.kt
@@ -5,7 +5,6 @@ import com.r8n.backend.core.api.PageResponseDto
 import com.r8n.backend.opinions.api.lists.dto.OpinionListDto
 import com.r8n.backend.opinions.integration.api.OpinionListsInternalApi
 import com.r8n.backend.opinions.integration.api.OpinionListsInternalApi.Companion.MINE_FULL_PATH
-import com.r8n.backend.opinions.integration.api.OpinionsInternalApi.Companion.RESTORE_PATH
 import org.springframework.web.client.RestClient
 import org.springframework.web.client.body
 
@@ -27,13 +26,4 @@ class OpinionListInternalRestClient(
                     }.build()
             }.retrieve()
             .body<PageResponseDto<OpinionListDto>>()!!
-
-    override fun restoreOpinionList(list: OpinionListDto) {
-        restClient
-            .post()
-            .uri(RESTORE_PATH)
-            .body(list)
-            .retrieve()
-            .toBodilessEntity()
-    }
 }
