+++ backend/opinions-sv/service/src/main/kotlin/com/r8n/backend/opinions/lists/controllerInternal/OpinionListInternalController.kt
@@ -17,9 +17,4 @@ class OpinionListInternalController(
     @PreAuthorize(Authority.IS_USER_OR_SERVICE)
     override fun getMineFull(pageable: PageRequestDto): PageResponseDto<OpinionListDto> =
         opinionListFacade.getListsFull(CurrentUserIdentifier.getCurrentUserId(), pageable)
-
-    @PreAuthorize(Authority.IS_SERVICE)
-    override fun restoreOpinionList(list: OpinionListDto) {
-        opinionListFacade.restoreOpinionList(list)
-    }
 }
