+++ backend/opinions-sv/service/src/main/kotlin/com/r8n/backend/opinions/opinions/service/OpinionService.kt
@@ -2,9 +2,9 @@ package com.r8n.backend.opinions.opinions.service
 
 import com.r8n.backend.opinions.access.domain.OpinionPermissionEnum
 import com.r8n.backend.opinions.access.service.AccessService
-import com.r8n.backend.opinions.opinions.database.ModerationDecisionRepository
 import com.r8n.backend.opinions.api.opinions.dto.OpinionDto
 import com.r8n.backend.opinions.api.opinions.dto.OpinionStatusEnumDto
+import com.r8n.backend.opinions.opinions.database.ModerationDecisionRepository
 import com.r8n.backend.opinions.opinions.database.OpinionRepository
 import com.r8n.backend.opinions.opinions.domain.ModerationDecision
 import com.r8n.backend.opinions.opinions.domain.ModerationDecisionAction
