+++ backend/users-sv/service/src/main/kotlin/com/r8n/backend/users/persistence/UserSessionPersistence.kt
@@ -2,7 +2,6 @@ package com.r8n.backend.users.persistence
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
 import jakarta.persistence.Id
 import jakarta.persistence.Table
 import org.hibernate.annotations.UuidGenerator
