+++ backend/users-sv/service/src/test/kotlin/com/r8n/backend/users/RestoreUserVerificationTest.kt
@@ -5,12 +5,13 @@ import com.r8n.backend.opinions.api.access.OutgoingAccessRequestApi
 import com.r8n.backend.opinions.integration.api.OpinionListsInternalApi
 import com.r8n.backend.security.ServiceTokenService
 import com.r8n.backend.users.api.dto.UserStatusEnumDto
+import com.r8n.backend.users.domain.UserStatusEnum
 import com.r8n.backend.users.integration.api.dto.UserDto
 import com.r8n.backend.users.persistence.PIIPersistence
 import com.r8n.backend.users.persistence.UserPersistence
-import com.r8n.backend.users.domain.UserStatusEnum
 import com.r8n.backend.users.provider.database.PIIRepository
 import com.r8n.backend.users.provider.database.UserRepository
+import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
@@ -26,10 +27,9 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
 import org.testcontainers.postgresql.PostgreSQLContainer
+import tools.jackson.databind.json.JsonMapper
 import java.time.Instant
 import java.util.UUID
-import tools.jackson.databind.json.JsonMapper
-import org.junit.jupiter.api.Assertions.assertEquals
 
 @ActiveProfiles("test")
 @Testcontainers
@@ -48,11 +48,12 @@ class RestoreUserVerificationTest {
         @Suppress("unused")
         @Container
         @ServiceConnection
-        val postgres = PostgreSQLContainer("postgres:15")
-            .withDatabaseName("users")
-            .withUsername("test")
-            .withPassword("test")
-            .withInitScript("db/init-schema.sql")
+        val postgres =
+            PostgreSQLContainer("postgres:15")
+                .withDatabaseName("users")
+                .withUsername("test")
+                .withPassword("test")
+                .withInitScript("db/init-schema.sql")
     }
 
     @Autowired
@@ -75,54 +76,58 @@ class RestoreUserVerificationTest {
         // Arrange: Create an existing user
         val existingUserId = UUID.randomUUID()
         val email = "existing@example.com"
-        val existingUser = UserPersistence(
-            id = existingUserId,
-            status = UserStatusEnum.ACTIVE,
-            statusTimestamp = Instant.now(),
-            lastSeenAt = Instant.now(),
-            passwordHash = "hash"
-        )
+        val existingUser =
+            UserPersistence(
+                id = existingUserId,
+                status = UserStatusEnum.ACTIVE,
+                statusTimestamp = Instant.now(),
+                lastSeenAt = Instant.now(),
+                passwordHash = "hash",
+            )
         userRepository.save(existingUser)
-        
-        val pii = PIIPersistence(
-            userId = existingUserId,
-            name = "Original Name",
-            email = email,
-            phone = null,
-            about = null,
-            location = null
-        )
+
+        val pii =
+            PIIPersistence(
+                userId = existingUserId,
+                name = "Original Name",
+                email = email,
+                phone = null,
+                about = null,
+                location = null,
+            )
         piiRepository.save(pii)
 
         val token = serviceTokenService.generateServiceToken()
-        val userDto = UserDto(
-            name = "New Name Attempt", // Should be ignored if user exists according to requirements (keep id, name, email, password)
-            email = email,
-            status = UserStatusEnumDto.SUSPENDED,
-            statusTimestamp = Instant.now(),
-            consents = emptyList()
-        )
+        val userDto =
+            UserDto(
+                name = "New Name Attempt",
+                email = email,
+                status = UserStatusEnumDto.SUSPENDED,
+                statusTimestamp = Instant.now(),
+                consents = emptyList(),
+            )
 
         // Act
-        val result = mockMvc
-            .perform(
-                post("/api/internal/users/import")
-                    .header("Authorization", "Bearer $token")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(userDto))
-            ).andExpect(status().isOk)
-            .andReturn()
+        val result =
+            mockMvc
+                .perform(
+                    post("/api/internal/users/import")
+                        .header("Authorization", "Bearer $token")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(userDto)),
+                ).andExpect(status().isOk)
+                .andReturn()
 
         val returnedIdString = result.response.contentAsString.replace("\"", "")
         val returnedId = UUID.fromString(returnedIdString)
 
         // Assert
         assertEquals(existingUserId, returnedId)
-        
+
         val updatedUser = userRepository.findById(existingUserId).get()
         assertEquals(UserStatusEnum.SUSPENDED, updatedUser.status)
         assertEquals("hash", updatedUser.passwordHash)
-        
+
         val updatedPii = piiRepository.findById(existingUserId).get()
         assertEquals("Original Name", updatedPii.name)
         assertEquals(email, updatedPii.email)
